+++ ft_strstr.c
@@ -6,13 +6,13 @@
 /*   By: ralee <ralee@student.42.us.org>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/29 11:20:23 by ralee             #+#    #+#             */
-/*   Updated: 2017/12/15 11:55:34 by ralee            ###   ########.fr       */
+/*   Updated: 2017/12/17 12:05:54 by ralee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static void		looper(int *counter, char **placeholder, char **onebyone)
+static void		looper(size_t *counter, char **placeholder, char **onebyone)
 {
 	*counter = 0;
 	*placeholder = *onebyone;
@@ -23,7 +23,7 @@ char			*ft_strstr(char *big, char *little)
 {
 	char	*placeholder;
 	char	*onebyone;
-	int		counter;
+	size_t		counter;
 
 	counter = 0;
 	placeholder = big;
